refactor: screensets stores typing simplified
