+++ src/screensets/chat-copy/chatCopyScreenset.tsx
@@ -6,7 +6,7 @@
 import { type ScreensetConfig, ScreensetCategory, uikitRegistry, registerSlice, I18nRegistry, Language, apiRegistry, screensetRegistry } from '@hai3/uicore';
 import { CHAT_COPY_SCREENSET_ID, CHAT_COPY_SCREEN_ID } from './ids';
 import { MessageSquareIcon, MESSAGE_SQUARE_COPY_ICON_ID } from './uikit/icons/MessageSquareIcon';
-import { chatCopyReducer, initializeChatCopyEffects, ChatCopyStateKeys } from './chatCopyStore';
+import chatCopyReducer, { initializeChatCopyEffects } from './slices/chatCopySlice';
 
 // Import for side effect - register API service
 import { CHAT_DOMAIN } from './api/ChatApiService';
@@ -72,7 +72,7 @@ uikitRegistry.registerIcons({
  * Register chat copy slice dynamically with uicore store
  * Screensets can add their state without modifying uicore package
  */
-registerSlice(ChatCopyStateKeys.State, chatCopyReducer, initializeChatCopyEffects);
+registerSlice(CHAT_COPY_SCREENSET_ID, chatCopyReducer, initializeChatCopyEffects);
 
 /**
  * Chat Copy Screenset Configuration
